Fix: move token retrieval to @After; add user cleanup in tearDown

+++ src/test/java/tests/CreateUserTest.java
@@ -1,19 +1,18 @@
-package ru.yandex.praktikum.api.tests;
+package tests;
 
+import clients.UserClient;
+import models.UserModel;
+import generators.UserGenerator;
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
-import org.apache.http.HttpStatus;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import ru.yandex.praktikum.api.clients.UserClient;
-import ru.yandex.praktikum.api.generators.UserGenerator;
-import ru.yandex.praktikum.api.models.UserModel;
 
+import static org.apache.http.HttpStatus.*;
 import static org.hamcrest.Matchers.*;
 
 public class CreateUserTest {
-
     private UserClient userClient;
     private UserModel user;
     private String accessToken;
@@ -21,16 +20,13 @@ public class CreateUserTest {
     @Before
     public void setUp() {
         userClient = new UserClient();
-        user = UserGenerator.getRandomUser();
     }
 
     @After
     public void tearDown() {
-        if (accessToken != null) {
-            userClient.deleteUser(accessToken);
-        } else if (user != null && user.getPassword() != null) {
+        if (user != null && user.getEmail() != null && user.getPassword() != null) {
             var loginResp = userClient.loginUser(user);
-            if (loginResp.statusCode() == HttpStatus.SC_OK) {
+            if (loginResp.statusCode() == SC_OK) {
                 accessToken = loginResp.then().extract().path("accessToken");
                 userClient.deleteUser(accessToken);
             }
@@ -39,64 +35,62 @@ public class CreateUserTest {
 
     @Test
     @DisplayName("Создание уникального пользователя")
-    @Description("Позитивный сценарий регистрации нового пользователя с валидными email, паролем и именем. Ожидается статус 200, наличие токена и корректные данные в ответе.")
+    @Description("Позитивный сценарий: регистрация с новыми данными")
     public void createUniqueUserTest() {
+        user = UserGenerator.getRandomUser();
         var response = userClient.createUser(user);
         response.then()
-                .statusCode(HttpStatus.SC_OK)
+                .statusCode(SC_OK)
                 .body("success", equalTo(true))
                 .body("user.email", equalTo(user.getEmail()))
                 .body("user.name", equalTo(user.getName()))
                 .body("accessToken", notNullValue());
-        accessToken = response.then().extract().path("accessToken");
     }
 
     @Test
     @DisplayName("Создание уже существующего пользователя")
-    @Description("Негативный сценарий: попытка зарегистрировать пользователя с уже существующими данными. Ожидается статус 403 и сообщение 'User already exists'.")
+    @Description("Негативный сценарий: повторная регистрация")
     public void createExistingUserTest() {
-        userClient.createUser(user).then().statusCode(HttpStatus.SC_OK);
+        user = UserGenerator.getRandomUser();
+        userClient.createUser(user).then().statusCode(SC_OK);
         var response = userClient.createUser(user);
         response.then()
-                .statusCode(HttpStatus.SC_FORBIDDEN)
+                .statusCode(SC_FORBIDDEN)
                 .body("success", equalTo(false))
                 .body("message", equalTo("User already exists"));
-        var loginResp = userClient.loginUser(user);
-        accessToken = loginResp.then().extract().path("accessToken");
     }
 
     @Test
     @DisplayName("Создание пользователя без пароля")
-    @Description("Негативный сценарий: отсутствует обязательное поле 'password'. Ожидается статус 403 и сообщение 'Email, password and name are required fields'.")
     public void createUserWithoutPasswordTest() {
+        user = UserGenerator.getRandomUser();
         user.setPassword(null);
         var response = userClient.createUser(user);
         response.then()
-                .statusCode(HttpStatus.SC_FORBIDDEN)
+                .statusCode(SC_FORBIDDEN)
                 .body("success", equalTo(false))
                 .body("message", equalTo("Email, password and name are required fields"));
-        accessToken = null; // пользователь не создан
     }
 
     @Test
     @DisplayName("Создание пользователя без имени")
-    @Description("Негативный сценарий: отсутствует поле 'name'. Ожидается статус 403 и соответствующее сообщение.")
     public void createUserWithoutNameTest() {
+        user = UserGenerator.getRandomUser();
         user.setName(null);
         var response = userClient.createUser(user);
         response.then()
-                .statusCode(HttpStatus.SC_FORBIDDEN)
+                .statusCode(SC_FORBIDDEN)
                 .body("message", equalTo("Email, password and name are required fields"));
     }
 
     @Test
     @DisplayName("Создание пользователя без email")
-    @Description("Негативный сценарий: отсутствует поле 'email'. Ожидается статус 403 и сообщение об обязательных полях.")
     public void createUserWithoutEmailTest() {
+        user = UserGenerator.getRandomUser();
         user.setEmail(null);
         var response = userClient.createUser(user);
         response.then()
-                .statusCode(HttpStatus.SC_FORBIDDEN)
+                .statusCode(SC_FORBIDDEN)
                 .body("message", equalTo("Email, password and name are required fields"));
     }
 }
